[fix] Reduced severity of output message
Inside the raster loading code, if certain values were not available
the code would print critical messages. I have changed that to warnings
to properly convey the issue as there are safe defaults that are applied
if they are missing. A warning is much more appropriate than a critical
messages